Added check if Extension should be active or not

+++ cogs/Channel.py
@@ -1,4 +1,5 @@
 import asyncio
+import json
 import discord
 from discord.ext import commands
 
@@ -16,21 +17,29 @@ class Channel(commands.Cog):
         Delete a textchannel
         - **?delchannel [`channel`]**
         """
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
 
-        if channel is not None:
-            await channel.delete()
-            await ctx.send(f"Der Channel ``{channel.name}`` wurde **Erfolgreich gelöscht!**", delete_after=5)
-            await ctx.message.delete()
-
-        elif not channel:
-            channel = ctx.channel
-            await ctx.channel.delete()
-            await ctx.send(f"Der Channel ``{channel.name}`` wurde **Erfolgreich gelöscht!**", delete_after=5)
-            await ctx.message.delete()
-
+        if data[str(ctx.guild.id)]["Channel"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Channel) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
         else:
-            await ctx.send(f'No channel named, "{channel}", was found', delete_after=5)
-            await ctx.message.delete()
+            if channel is not None:
+                await channel.delete()
+                await ctx.send(f"Der Channel ``{channel.name}`` wurde **Erfolgreich gelöscht!**", delete_after=5)
+                await ctx.message.delete()
+
+            elif not channel:
+                channel = ctx.channel
+                await ctx.channel.delete()
+                await ctx.send(f"Der Channel ``{channel.name}`` wurde **Erfolgreich gelöscht!**", delete_after=5)
+                await ctx.message.delete()
+
+            else:
+                await ctx.send(f'No channel named, "{channel}", was found', delete_after=5)
+                await ctx.message.delete()
 
     @commands.command(name='createchannel', aliases=['mkchannel'])
     async def create_channel(self, ctx, *, channel_name):
@@ -39,20 +48,26 @@ class Channel(commands.Cog):
         - **?mkchannel [`channel_name`]**
         """
 
-        guild = ctx.message.guild
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
 
-        if not channel_name:
-            await ctx.send('Du hast **keinen Channel Namen angegeben!**', delete_after=5)
-            await ctx.message.delete()
-
-        if channel_name is not None:
-            channel = await guild.create_text_channel(f'{channel_name}')
-            await ctx.send(f'Der Channel {channel.mention} wurde **Erfolgreich erstellt!**', delete_after=5)
-            await ctx.message.delete()
-
+        if data[str(ctx.guild.id)]["Channel"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Channel) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
         else:
-            await ctx.send(f'Irgendetwas ist schiefgelaufen!', delete_after=5)
-            await ctx.message.delete()
+            guild = ctx.message.guild
+            if not channel_name:
+                await ctx.send('Du hast **keinen Channel Namen angegeben!**', delete_after=5)
+                await ctx.message.delete()
+            if channel_name is not None:
+                channel = await guild.create_text_channel(f'{channel_name}')
+                await ctx.send(f'Der Channel {channel.mention} wurde **Erfolgreich erstellt!**', delete_after=5)
+                await ctx.message.delete()
+            else:
+                await ctx.send(f'Irgendetwas ist schiefgelaufen!', delete_after=5)
+                await ctx.message.delete()
 
     @commands.command(name='tempchannel', aliases=['temp_channel'])
     @commands.cooldown(rate=5, per=1)
@@ -69,22 +84,28 @@ class Channel(commands.Cog):
         "w": week/s`
         """
 
-        if not channel_name:
-            await ctx.send('Du hast keine Channel Name angegeben!', delete_after=5)
+        with open('utils/json/active_check.json', 'r') as f:
+            data = json.load(f)
 
-        if not time:
-            await ctx.send('Du hast keine Zeit angegeben!', delete_after=5)
-
-        guild = ctx.message.guild
-        time_convert = {"s": 1, "m": 60, "h": 3600, "d": 86400, "w": 604800}
-        tempchanneltime = int(time[:-1]) * time_convert[time[-1]]
-
-        await ctx.message.delete()
-        channel = await guild.create_text_channel(f'{channel_name}')
-        await ctx.send(f'Temp Channel `{channel_name}` **Erfolgreich erstellt!**', delete_after=5)
-        await asyncio.sleep(tempchanneltime)
-        await channel.delete()
-        await ctx.send(f'Temp Channel `{channel_name}` **Erfolgreich gelöscht!**', delete_after=5)
+        if data[str(ctx.guild.id)]["Channel"] == 'false':
+            embed = discord.Embed(
+                description=f'Diese **Extension (Channel) ist momentan deaktiviert!** Wende dich bitte an **den Owner vom Bot** (LookAtYourSkill#6666)',
+                color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            if not channel_name:
+                await ctx.send('Du hast keine Channel Name angegeben!', delete_after=5)
+            if not time:
+                await ctx.send('Du hast keine Zeit angegeben!', delete_after=5)
+            guild = ctx.message.guild
+            time_convert = {"s": 1, "m": 60, "h": 3600, "d": 86400, "w": 604800}
+            tempchanneltime = int(time[:-1]) * time_convert[time[-1]]
+            await ctx.message.delete()
+            channel = await guild.create_text_channel(f'{channel_name}')
+            await ctx.send(f'Temp Channel `{channel_name}` **Erfolgreich erstellt!**', delete_after=5)
+            await asyncio.sleep(tempchanneltime)
+            await channel.delete()
+            await ctx.send(f'Temp Channel `{channel_name}` **Erfolgreich gelöscht!**', delete_after=5)
 
 def setup(bot):
     bot.add_cog(Channel(bot))
